% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CimpleG.R
\name{CimpleG}
\alias{CimpleG}
\title{Main function of the package
Finds the best predictors (CpGs) for the given targets}
\usage{
CimpleG(
  train_data,
  test_data,
  train_targets,
  test_targets,
  targets,
  method = c("adhoc", "parab", "parab_scale", "oner"),
  k_folds = 10,
  n_repeats = 1
)
}
\description{
Main function of the package
Finds the best predictors (CpGs) for the given targets
}
