% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CimpleG.R
\name{CimpleG_deconvolution}
\alias{CimpleG_deconvolution}
\title{Perform deconvolution using the results from CimpleG
Calculate reference matrix given reference data and targets
Compute deconvolution on new data}
\usage{
CimpleG_deconvolution(
  CimpleG_result,
  reference_data,
  reference_targets,
  targets,
  new_data,
  method = c("nnls", "epidish", "nmf")
)
}
\description{
Perform deconvolution using the results from CimpleG
Calculate reference matrix given reference data and targets
Compute deconvolution on new data
}
