% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CimpleG.R
\name{CimpleG_general}
\alias{CimpleG_general}
\title{Main function of the package
Finds the best predictors (CpGs) for the given targets}
\usage{
CimpleG_general(
  train_data,
  test_data,
  train_targets,
  test_targets,
  targets,
  model_type = c("logistic_reg", "decision_tree", "boost_tree", "mlp", "rand_forest",
    "null_model"),
  engine = c("glmnet", "xgboost", "nnet", "ranger"),
  k_folds = 10,
  n_repeats = 1,
  grid_n = 10
)
}
\description{
Main function of the package
Finds the best predictors (CpGs) for the given targets
}
