% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{compute_ax}
\alias{compute_ax}
\title{Feature selection function used in the sigma delta space}
\usage{
compute_ax(dm, sv, p)
}
\arguments{
\item{dm, }{delta (difference in mean values)}

\item{sv, }{sigma (sum of variance values)}

\item{p, }{even number, the greater 'p' is the more importance will be given to sigma}
}
\description{
Feature selection function used in the sigma delta space
}
