% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{deconvolution_epidish}
\alias{deconvolution_epidish}
\title{EpiDISH deconvolution}
\usage{
deconvolution_epidish(
  ref_mat,
  new_data,
  epidish_method = "CBS",
  epidish_nuv = seq(0.1, 1, 0.1),
  epidish_maxit = 10000
)
}
\arguments{
\item{ref_mat}{Reference matrix.}

\item{new_data}{New data matrix.}

\item{epidish_method}{One of `CBS` (Cibersort), `RPC` (Robust Partial Correlations), `CP` (Constrained Projection).
Default is `CBS`. See `EpiDISH` documentation for more information.}

\item{epidish_nuv}{A vector of candidate values used for svm. Only used when epidish_method is set to `CBS`.
See `EpiDISH` documentation for more information.}

\item{epidish_maxit}{Integer with the number of max iterations for IWLS (Iterative Weighted Least Squares).
Only used when epidish_method is set to `RPC`.}
}
\description{
EpiDISH deconvolution
}
