% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{deconvolution_nmf}
\alias{deconvolution_nmf}
\title{NMF deconvolution}
\usage{
deconvolution_nmf(weights_mat, values_mat, ...)
}
\arguments{
\item{weights_mat}{Reference matrix.}

\item{values_mat}{New data matrix.}

\item{...}{Extra parameters to be set NMF options. Most relevant parameters are probably `method` and `beta`.}
}
\description{
NMF deconvolution
}
