% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{deconvolution_nnls}
\alias{deconvolution_nnls}
\title{NNLS deconvolution}
\usage{
deconvolution_nnls(dt, compute_cols, ref_mat)
}
\arguments{
\item{dt}{A data.table with the new data with features/predictions on rows and samples on columns.}

\item{compute_cols}{A character vector with the columns for which the deconvolution algorithm should be ran.}

\item{ref_mat}{The reference matrix as created by CimpleG.}
}
\description{
NNLS deconvolution
}
