% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{do_cv}
\alias{do_cv}
\title{Main function controlling k-fold CV training}
\usage{
do_cv(
  train_data,
  k_folds = 10,
  n_repeats = 1,
  method = c("parab", "parab_scale", "adhoc", "oner")
)
}
\description{
Main function controlling k-fold CV training
}
