% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{final_model}
\alias{final_model}
\title{Train the final model after training}
\usage{
final_model(train_data, best_pred, train_summary, method = "oner")
}
\description{
Train the final model after training
}
