% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{find_predictors}
\alias{find_predictors}
\title{Training/Finding the best predictors/CpGs}
\usage{
find_predictors(
  split_train_set,
  init_step = 0.1,
  step_increment = 0.1,
  min_feat_search = 10,
  p_class = "positive_class",
  method,
  scale_scores = FALSE,
  pred_type = c("both", "hypo", "hyper")
)
}
\description{
Training/Finding the best predictors/CpGs
}
