% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{fix_oner_boundaries}
\alias{fix_oner_boundaries}
\title{Fixes lower and upper boundaries (0,1) on OneR models}
\usage{
fix_oner_boundaries(oner_mod)
}
\value{
OneR object
}
\description{
Fixes lower and upper boundaries (0,1) on OneR models
}
