% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_load_helper.R
\name{load_object}
\alias{load_object}
\title{Load an R object saved with CimpleG or an RDS file.}
\usage{
load_object(file_name)
}
\arguments{
\item{file_name}{File name in the working directory or path to file
to be loaded. Files saved with \code{CimpleG::save_object} and \code{.rds}
files are supported. If the \code{archive} package is installed.}
}
\description{
Load an R object saved with CimpleG or an RDS file.
}
