% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{make_deconv_ref_matrix}
\alias{make_deconv_ref_matrix}
\title{Build deconvolution reference matrix}
\usage{
make_deconv_ref_matrix(cpg_obj, ref_data, ref_data_labels, method = NULL)
}
\arguments{
\item{cpg_obj}{A CimpleG object.}

\item{ref_data}{A matrix with the reference data to be used to build the reference matrix.}

\item{ref_data_labels}{A character vector with the true labels of the samples in the `reference_data`.}

\item{method}{Method used to train models in the CimpleG object. 
If not provided (NULL), method will be taken from the CimpleG object.}
}
\description{
Build deconvolution reference matrix
}
