% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{parabola_iter_loop}
\alias{parabola_iter_loop}
\title{Loop iterating through parabola selecting cpgs}
\usage{
parabola_iter_loop(
  df_diffmean_sumvar,
  init_step,
  step_increment,
  min_feat_search,
  pred_type
)
}
\value{
list
}
\description{
Loop iterating through parabola selecting cpgs
}
