% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_deconvolution.R
\name{plot_deconvolution}
\alias{plot_deconvolution}
\title{Deconvolution plotting functions}
\usage{
plot_deconvolution(
  deconv_mat,
  name_tag,
  sorted_classes,
  sort_by_class_value = TRUE,
  color_palette_df = NULL,
  facet_by_sample_group = FALSE,
  group_split_chr = "_",
  group_split_pos = 2,
  facet_group_order = NULL,
  subtitle = NULL,
  plot_dir = NULL,
  time_stamp = TRUE,
  hide_x_axis = TRUE,
  size_x_axis = NULL,
  facet_x_space = c("free_x", "fixed"),
  do_boxplots = FALSE,
  show_legend = TRUE,
  signif = TRUE
)
}
\arguments{
\item{deconv_mat}{TODO}

\item{name_tag}{TODO}

\item{sorted_classes}{TODO}

\item{sort_by_class_value}{TODO}

\item{color_palette_df}{TODO}

\item{facet_by_sample_group}{TODO}

\item{group_split_chr}{TODO}

\item{group_split_pos}{TODO}

\item{facet_group_order}{TODO}

\item{subtitle}{TODO}

\item{plot_dir}{TODO}

\item{time_stamp}{TODO}

\item{hide_x_axis}{TODO}

\item{size_x_axis}{TODO}

\item{facet_x_space}{TODO}

\item{do_boxplots}{TODO}

\item{show_legend}{TODO}

\item{signif}{TODO}
}
\description{
Deconvolution plotting functions
}
