% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_load_helper.R
\name{save_object}
\alias{save_object}
\title{Save an R object to disk with fast and efficient compression algorithms.}
\usage{
save_object(object, file_name, file_format = "zstd")
}
\arguments{
\item{object}{Object to be saved to disk.}

\item{file_name}{Name of the file where the R object is saved to.}

\item{file_format}{One of "zstd", "lz4", "gzip", "bzip2","xz", "nocomp".
\code{zstd} is the best option, fast compression and loading times, low space usage.
Formats "zstd" and "lz4" are only available if package \code{archive} is installed.}
}
\description{
Save an R object to disk with fast and efficient compression algorithms.
}
