% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{select_features}
\alias{select_features}
\title{Feature selection function used in the diffmeans, sumvariance space}
\usage{
select_features(x, y, a)
}
\arguments{
\item{x, }{difference in means value}

\item{y, }{sum of variances value}

\item{a, }{parabola parameter, scales how open/closed the parabola is, the higher the value, the more closed the parabola is.}
}
\value{
bool vector
}
\description{
Feature selection function used in the diffmeans, sumvariance space
}
