% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{Cell line test data}
\format{
A matrix with 1000 CpGs/features/variables as columns and 170 samples as rows
}
\usage{
test_data
}
\description{
Cell line test data
}
\keyword{datasets}
