% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{train_data}
\alias{train_data}
\title{Cell line train data}
\format{
A matrix with 1000 CpGs/features/variables as columns and 409 samples as rows
}
\usage{
train_data
}
\description{
Cell line train data
}
\keyword{datasets}
