% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{train_targets}
\alias{train_targets}
\title{Cell line train data targets}
\format{
A data frame with 18 variables for 409 samples as rows.
\describe{
\item{\code{gsm}}{GSM identifier (GEO accession number) of the sample}
\item{\code{cell_type}}{the cell type of the respective sample}
\item{\code{adipocytes}}{one-hot encoded (1 or 0) column defining if a given sample is an adipocyte}
\item{\code{astrocytes}}{one-hot encoded (1 or 0) column defining if a given sample is an astrocyte}
\item{\code{blood_cells}}{one-hot encoded (1 or 0) column defining if a given sample is a blood cell}
\item{\code{endothelial_cells}}{one-hot encoded (1 or 0) column defining if a given sample is an endothelial cell}
\item{\code{epidermal_cells}}{one-hot encoded (1 or 0) column defining if a given sample is an epidermal cell}
\item{\code{epithelial_cells}}{one-hot encoded (1 or 0) column defining if a given sample is an epithelial cell}
\item{\code{fibroblasts}}{one-hot encoded (1 or 0) column defining if a given sample is a fibroblast}
\item{\code{glia}}{one-hot encoded (1 or 0) column defining if a given sample is a glia cell}
\item{\code{hepatocytes}}{one-hot encoded (1 or 0) column defining if a given sample is an hepatocyte}
\item{\code{ips_cells}}{one-hot encoded (1 or 0) column defining if a given sample is an ipsc}
\item{\code{msc}}{one-hot encoded (1 or 0) column defining if a given sample is an msc}
\item{\code{muscle_cells}}{one-hot encoded (1 or 0) column defining if a given sample is a muscle cell}
\item{\code{neurons}}{one-hot encoded (1 or 0) column defining if a given sample is a neuron}
\item{\code{muscle_sc}}{one-hot encoded (1 or 0) column defining if a given sample is a muscle stem cell}
\item{\code{group_data}}{to which dataset these data belong to (\code{train} or \code{test})}
\item{\code{description}}{the cell type of the respective sample, in long form}
}
}
\usage{
train_targets
}
\description{
Cell line train data targets
}
\keyword{datasets}
