% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CimpleG-package.R
\docType{package}
\name{CimpleG-package}
\alias{CimpleG-package}
\title{CimpleG: A Package to Identify Single CpG Sites for Classification and Deconvolution}
\description{
DNA methylation signatures are usually based on multivariate approaches that require hundreds of sites for predictions. CimpleG is a method for the detection of small CpG methylation signatures used for cell-type classification and deconvolution. CimpleG is time efficient and performs as well as top performing methods for cell-type classification of blood cells and other somatic cells, while basing its prediction on a single DNA methylation site per cell type (but users can also select more sites if they so wish). Users can train cell type classifiers (CimpleG based, and others) and directly apply these in a deconvolution of cell mixes context. Altogether, CimpleG provides a complete computational framework for the delineation of DNAm signatures and cellular deconvolution. For more details see Maié et al. (2023) \doi{10.1186/s13059-023-03000-0}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CostaLab/CimpleG}
  \item \url{https://costalab.github.io/CimpleG/}
  \item Report bugs at \url{https://github.com/CostaLab/CimpleG/issues}
}

}
\author{
\strong{Maintainer}: Tiago F.V. Maié \email{tiagomaie@hotmail.com}

}
\keyword{internal}
