% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CimpleG-package.R
\docType{package}
\name{CimpleG-package}
\alias{CimpleG-package}
\alias{_PACKAGE}
\title{CimpleG: CimpleG - Individual/Single CpG sites For Classification and Deconvolution}
\description{
CimpleG is a simple package to find single-CpG based signatures for use in classification and deconvolution.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CostaLab/CimpleG}
  \item \url{https://costalab.github.io/CimpleG/}
  \item Report bugs at \url{https://github.com/CostaLab/CimpleG/issues}
}

}
\author{
\strong{Maintainer}: Tiago F.V. Maié \email{tiagomaie@hotmail.com}

}
\keyword{internal}
