% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CimpleG.R
\name{CimpleG}
\alias{CimpleG}
\title{Method to find the best simple predictors (CpGs) for the given target data.}
\usage{
CimpleG(
  train_data,
  train_targets,
  targets,
  test_data = NULL,
  test_targets = NULL,
  method = c("parab_scale", "adhoc", "parab", "oner"),
  pred_type = c("both", "hypo", "hyper"),
  k_folds = 10,
  n_repeats = 1,
  n_cores = 1
)
}
\arguments{
\item{train_data}{Training dataset.
A matrix (s x f) with methylation data (Beta values)
that will be used to train/find the predictors.
Samples (s) must be in rows while features/CpGs (f) must be in columns.}

\item{train_targets}{A data frame with the training target samples one-hot encoded.
A data frame with at least 1 column,
with as many rows and in the same order as `train_data`.
Target columms need to be one-hot encoded, meaning that, for that column
the target samples should be encoded as `1` while every other sample
should be encoded as `0`.}

\item{targets}{A string specifying the name of the column in `train_targets`
to be used for training.
Can be a character vector if there are several columns in `train_targets`
to be used for training.
If this argument is a character vector, CimpleG will search for the
best predictors for each target sequentially or in parallel depending on the
value of `n_cores`}

\item{test_data}{Testing dataset.
A matrix (s x f) with methylation data (Beta values)
that will be used to test the performance of the found predictors.
Samples (s) must be in rows while features/CpGs (f) must be in columns.
If `test_data` *OR* `test_targets` are NULL, CimpleG will generate a
stratified test dataset based on `train_targets` by removing 25% of the
samples from `train_data` and `train_targets`.}

\item{test_targets}{A data frame with the testing target samples one-hot encoded.
A data frame with at least 1 column,
with as many rows and in the same order as `test_data`.
Target columms need to be one-hot encoded, meaning that, for that column
the target samples should be encoded as `1` while every other sample
should be encoded as `0`.
If `test_data` *OR* `test_targets` are NULL, CimpleG will generate a
stratified test dataset based on `train_targets` by removing 25% of the
samples from `train_data` and `train_targets`.}

\item{method}{A string specifying the method to be used for training.
One of `parab_scale` (default), `adhoc`, `parab` or `oner`}

\item{pred_type}{A string specifying the type of predictor/CpG to be
searched for during training.
One of `both` (default), `hypo` or `hyper`.
If `hypo`, only hypomethylated predictors will be considered.
If `hyper`, only hypermethylated predictors will be considered.}

\item{k_folds}{An integer specifying the number of folds (K) to be used
in training for the stratified K-fold cross-validation procedure.}

\item{n_repeats}{An integer specifying the number of repeats (N) to be used
in training for the stratified N-repeats K-fold cross-validation procedure.
Currently not in use.}

\item{n_cores}{If larger than `1`, the default, it will use that many cores,
to search for predictors for multiple targets (up to `n_cores`)
at the same time (parallel processing) in order to save
in computational time.}
}
\description{
Method to find the best simple predictors (CpGs) for the given target data.
}
\examples{
library("CimpleG")

# read data
train_d <- readRDS(file.path("data","mock_train.RDS"))
test_d <- readRDS(file.path("data","mock_test.RDS"))
train_target <- readRDS(file.path("data","mock_train_targets.RDS"))
test_target <- readRDS(file.path("data","mock_test_targets.RDS"))

# run CimpleG
cimpleg_result <- CimpleG(
  train_data = train_d,
  train_targets = train_target,
  test_data = test_d,
  test_targets = test_target,
  targets = c("CELL_TYPE_MSCORFIBRO","CELL_TYPE_NEURONS")
)

# check results
cimpleg_result$results
cimpleg_result$signatures

}
