% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CimpleG.R
\name{CimpleG}
\alias{CimpleG}
\title{Train a classification model using (CpGs) as features
for the given target data.}
\usage{
CimpleG(
  train_data,
  train_targets,
  targets,
  test_data = NULL,
  test_targets = NULL,
  method = c("parab_scale", "adhoc", "parab", "oner", "logistic_reg", "decision_tree",
    "boost_tree", "mlp", "rand_forest", "null_model"),
  pred_type = c("both", "hypo", "hyper"),
  engine = c("glmnet", "xgboost", "nnet", "ranger"),
  k_folds = 10,
  grid_n = 10,
  n_repeats = 1,
  run_parallel = FALSE
)
}
\arguments{
\item{train_data}{Training dataset.
A matrix (s x f) with methylation data (Beta values)
that will be used to train/find the predictors.
Samples (s) must be in rows while features/CpGs (f) must be in columns.}

\item{train_targets}{A data frame with the training target samples one-hot encoded.
A data frame with at least 1 column,
with as many rows and in the same order as `train_data`.
Target columms need to be one-hot encoded, meaning that, for that column
the target samples should be encoded as `1` while every other sample
should be encoded as `0`.}

\item{targets}{A string specifying the name of the column in `train_targets`
to be used for training.
Can be a character vector if there are several columns in `train_targets`
to be used for training.
If this argument is a character vector, CimpleG will search for the
best predictors for each target sequentially or in parallel depending on the
value of `run_parallel`}

\item{test_data}{Testing dataset.
A matrix (s x f) with methylation data (Beta values)
that will be used to test the performance of the found predictors.
Samples (s) must be in rows while features/CpGs (f) must be in columns.
If `test_data` *OR* `test_targets` are NULL, CimpleG will generate a
stratified test dataset based on `train_targets` by removing 25% of the
samples from `train_data` and `train_targets`.}

\item{test_targets}{A data frame with the testing target samples one-hot encoded.
A data frame with at least 1 column,
with as many rows and in the same order as `test_data`.
Target columms need to be one-hot encoded, meaning that, for that column
the target samples should be encoded as `1` while every other sample
should be encoded as `0`.
If `test_data` *OR* `test_targets` are NULL, CimpleG will generate a
stratified test dataset based on `train_targets` by removing 25% of the
samples from `train_data` and `train_targets`.}

\item{method}{A string specifying the method or type of machine learning
model/algorithm to be used for training.
These are divided in two main groups.
* The simple models (classifiers that use a single feature),
`parab_scale` (default), `adhoc`, `parab` or `oner`;
* the complex models (classifiers that use several features),
`logistic_reg`, `decision_tree`, `boost_tree`, `mlp` or `rand_forest`.}

\item{pred_type}{A string specifying the type of predictor/CpG to be
searched for during training. Only used for simple models.
One of `both` (default), `hypo` or `hyper`.
If `hypo`, only hypomethylated predictors will be considered.
If `hyper`, only hypermethylated predictors will be considered.}

\item{engine}{A string specifying the
machine learning engine behind `method`. Only used for complex models.
Currently not in use.}

\item{k_folds}{An integer specifying the number of folds (K) to be used
in training for the stratified K-fold cross-validation procedure.}

\item{grid_n}{An integer specifying the number of hyperparameter combinations
to train for.}

\item{n_repeats}{An integer specifying the number of repeats (N) to be used
in training for the stratified N-repeats K-fold cross-validation procedure.
Currently not in use.}

\item{run_parallel}{A boolean, if `FALSE`, the default, it will search
for predictors for multiple targets sequentially.
If `TRUE` it will search for predictors for multiple targets
at the same time (parallel processing) in order to save
in computational time. You need to set up `future::plan()` before running
this function.}
}
\description{
Train a classification model using (CpGs) as features
for the given target data.
}
\examples{
library("CimpleG")

# read data
data(train_data)
data(train_targets)
data(test_data)
data(test_targets)

# run CimpleG
cimpleg_result <- CimpleG(
  train_data = train_data,
  train_targets = train_targets,
  test_data = test_data,
  test_targets = test_targets,
  method = "parab_scale",
  targets = c("CELL_TYPE_MSCORFIBRO","CELL_TYPE_NEURONS")
)

# check results
cimpleg_result$results

}
