% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CimpleG.R
\name{CimpleG_general}
\alias{CimpleG_general}
\title{Train a classification model using (CpGs) as features
for the given target data.}
\usage{
CimpleG_general(
  train_data,
  train_targets,
  targets,
  test_data = NULL,
  test_targets = NULL,
  model_type = c("logistic_reg", "decision_tree", "boost_tree", "mlp", "rand_forest",
    "null_model"),
  engine = c("glmnet", "xgboost", "nnet", "ranger"),
  k_folds = 10,
  n_repeats = 1,
  grid_n = 10,
  n_cores = 1
)
}
\arguments{
\item{train_data}{Training dataset.
A matrix (s x f) with methylation data (Beta values)
that will be used to train/find the predictors.
Samples (s) must be in rows while features/CpGs (f) must be in columns.}

\item{train_targets}{A data frame with the training target samples one-hot encoded.
A data frame with at least 1 column,
with as many rows and in the same order as `train_data`.
Target columms need to be one-hot encoded, meaning that, for that column
the target samples should be encoded as `1` while every other sample
should be encoded as `0`.}

\item{targets}{A string specifying the name of the column in `train_targets`
to be used for training.
Can be a character vector if there are several columns in `train_targets`
to be used for training.
If this argument is a character vector, CimpleG will search for the
best predictors for each target sequentially or in parallel depending on the
value of `n_cores`}

\item{test_data}{Testing dataset.
A matrix (s x f) with methylation data (Beta values)
that will be used to test the performance of the found predictors.
Samples (s) must be in rows while features/CpGs (f) must be in columns.
If `test_data` *OR* `test_targets` are NULL, CimpleG will generate a
stratified test dataset based on `train_targets` by removing 25% of the
samples from `train_data` and `train_targets`.}

\item{test_targets}{A data frame with the testing target samples one-hot encoded.
A data frame with at least 1 column,
with as many rows and in the same order as `test_data`.
Target columms need to be one-hot encoded, meaning that, for that column
the target samples should be encoded as `1` while every other sample
should be encoded as `0`.
If `test_data` *OR* `test_targets` are NULL, CimpleG will generate a
stratified test dataset based on `train_targets` by removing 25% of the
samples from `train_data` and `train_targets`.}

\item{model_type}{A string specifying the type of
machine learning model/algorithm to be used for training.
One of `logistic_reg` (default), `decision_tree`,
`boost_tree`, `mlp` or `rand_forest`.}

\item{engine}{A string specifying the
machine learning engine behind `model_type`.
Currently not in use.}

\item{k_folds}{An integer specifying the number of folds (K) to be used
in training for the stratified K-fold cross-validation procedure.}

\item{n_repeats}{An integer specifying the number of repeats (N) to be used
in training for the stratified N-repeats K-fold cross-validation procedure.
Currently not in use.}

\item{n_cores}{If larger than `1`, the default, it will use that many cores,
to search for predictors for multiple targets (up to `n_cores`)
at the same time (parallel processing) in order to save
in computational time.}
}
\description{
Train a classification model using (CpGs) as features
for the given target data.
}
