% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{compute_diffmeans_sumvar}
\alias{compute_diffmeans_sumvar}
\title{Compute diff mean sum var dataframe}
\usage{
compute_diffmeans_sumvar(data, target_vector)
}
\arguments{
\item{data}{Matrix with beta values that will be used to compute diffmeans sumvar data frame}

\item{target_vector}{boolean vector defining which samples in data are part of the target class}
}
\description{
Compute diff mean sum var dataframe
}
