% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{darken}
\alias{darken}
\title{Helper function to darken down a given color.}
\usage{
darken(color, factor = 0.5)
}
\arguments{
\item{color}{Color name or hex code of a color}

\item{factor}{Multiplicative factor by which `color` will be darkened down}
}
\description{
Helper function to darken down a given color.
}
