% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{diffmeans_sumvariance_plot}
\alias{diffmeans_sumvariance_plot}
\title{Represent CpGs in the difference in means, sum of variances space.
This plot is often used to select CpGs that would be good classifiers.
These CpGs are often located on the bottom left and
bottom right of this plot.}
\usage{
diffmeans_sumvariance_plot(
  data,
  xcol = "diff_means",
  ycol = "sum_variance",
  feature_id_col = "id",
  is_feature_selected_col = NULL,
  label_var1 = "Target",
  label_var2 = "Others",
  target_vector = NULL,
  mean_cutoff = NULL,
  var_cutoff = NULL,
  threshold_func = NULL,
  func_factor = NULL,
  feats_to_highlight = NULL,
  cpg_ranking_df = NULL,
  color_all_points = NULL,
  pltDensity = TRUE,
  density_type = c("density", "histogram", "boxplot", "violin", "densigram"),
  plot_dir = NULL,
  id_tag = NULL,
  file_tag = NULL,
  custom_mods = FALSE
)
}
\arguments{
\item{data}{Data to create difference in means, sum of variances plot.
Either a data.frame with `xcol`,`ycol` and `feature_id_col` or, if
`target_vector` is not `NULL` a matrix with beta values from which,
given the target, the difference in means between the target and others,
and the sum of variances within the target and others will be calculated.}
}
\description{
Represent CpGs in the difference in means, sum of variances space.
This plot is often used to select CpGs that would be good classifiers.
These CpGs are often located on the bottom left and
bottom right of this plot.
}
