% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{lighten}
\alias{lighten}
\title{Helper function to lighten up a given color.}
\usage{
lighten(color, factor = 0.5)
}
\arguments{
\item{color}{Color name or hex code of a color}

\item{factor}{Multiplicative factor by which `color` will be lightened up}
}
\description{
Helper function to lighten up a given color.
}
