% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_plot_functions.R
\name{make_color_palette}
\alias{make_color_palette}
\title{Make color palette data frame}
\usage{
make_color_palette(classes)
}
\arguments{
\item{classes}{Vector with classes for which to create a color palette}
}
\description{
Make color palette data frame
}
