% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_deconvolution.R
\name{plot_deconvolution}
\alias{plot_deconvolution}
\title{Deconvolution plotting functions}
\usage{
plot_deconvolution(
  deconv_mat,
  name_tag,
  sorted_classes = c("Epithelial cells", "Hepatocytes", "Glia", "Neurons",
    "Endothelial cells", "Induced pluripotent stem cells", "Blood cells", "MSCorFibro"),
  sort_by_class_value = TRUE,
  color_palette_df = NULL,
  facet_by_sample_group = FALSE,
  group_split_chr = "_",
  group_split_pos = 2,
  facet_group_order = NULL,
  subtitle = NULL,
  plot_dir = NULL,
  time_stamp = TRUE,
  hide_x_axis = TRUE,
  size_x_axis = NULL,
  facet_x_space = c("free_x", "fixed"),
  do_boxplots = FALSE,
  show_legend = TRUE,
  signif = TRUE
)
}
\description{
Deconvolution plotting functions
}
