% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{predict.CimpleG}
\alias{predict.CimpleG}
\title{Predict outcome from a CimpleG signatures on new data}
\usage{
\method{predict}{CimpleG}(object, new_data, class_labels = NULL)
}
\arguments{
\item{object}{CimpleG object.}

\item{new_data}{Data to be predicted, samples should be in rows and features in columns.
Last column of `new_data` should have the target/class labels coded as 0 or 1.}

\item{class_labels}{Class labels of new data if these are not provided directly with it.}
}
\description{
Predict outcome from a CimpleG signatures on new data
}
